This commit reorganizes TOGA into an installable Python package, and adds MCNP MGXS output format conversion features
