+++ .gitignore
@@ -1,0 +1,246 @@
+# Compiled objects and modules
+*.a
+*.o
+*.mod
+*.log
+*.out
+
+# Compiler python objects
+*.pyc
+
+# Python distribution
+.settings/
+dist/
+openmc.egg-info/
+
+# Inputs generated from Python API
+examples/**/*.xml
+
+# emacs and vim backups
+*~
+*.swp
+
+# OpenMC statepoints
+*.binary
+
+# Documentation builds
+docs/build
+docs/source/_images/*.pdf
+docs/source/_images/*.aux
+docs/source/pythonapi/generated/
+
+# Source build
+build
+
+# build from src/utils/setup.py
+src/utils/build
+
+# xml-fortran reader
+src/xml-fortran/xmlreader
+
+# Test results error file
+results_error.dat
+inputs_error.dat
+results_test.dat
+
+# Test
+.pytest_cache/
+
+# HDF5 files
+*.h5
+
+
+# Nuclear data
+scripts/nndc
+scripts/nndc_hdf5
+scripts/wmp
+scripts/JEFF32-ACE-*.zip
+scripts/TSLs.tar.gz
+scripts/jeff-3.2
+scripts/jeff-3.2-hdf5
+scripts/*.tar.*
+scripts/G4EMLOW*/
+
+# Images
+*.ppm
+*.voxel
+*.vti
+
+# PyCharm project configuration files
+.idea
+.idea/*
+
+# IPython notebook checkpoints
+.ipynb_checkpoints
+
+
+# Cython files
+*.c
+*.html
+*.so
+
+.cache/
+.tox/
+.python-version
+.coverage
+htmlcov
+
+# macOS
+*.DS_Store
+
+# Dynamic Library
+*.dylib
+*.lib
+*.dll
+
+
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
+
+# C extensions
+*.so
+
+# Distribution / packaging
+.Python
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+.eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+wheels/
+share/python-wheels/
+*.egg-info/
+.installed.cfg
+*.egg
+MANIFEST
+
+# PyInstaller
+#  Usually these files are written by a python script from a template
+#  before PyInstaller builds the exe, so as to inject date/other infos into it.
+*.manifest
+*.spec
+
+# Installer logs
+pip-log.txt
+pip-delete-this-directory.txt
+
+# Unit test / coverage reports
+htmlcov/
+.tox/
+.nox/
+.coverage
+.coverage.*
+.cache
+nosetests.xml
+coverage.xml
+*.cover
+*.py,cover
+.hypothesis/
+.pytest_cache/
+cover/
+
+# Translations
+*.mo
+*.pot
+
+# Django stuff:
+*.log
+local_settings.py
+db.sqlite3
+db.sqlite3-journal
+
+# Flask stuff:
+instance/
+.webassets-cache
+
+# Scrapy stuff:
+.scrapy
+
+# Sphinx documentation
+docs/_build/
+
+# PyBuilder
+.pybuilder/
+target/
+
+# Jupyter Notebook
+.ipynb_checkpoints
+
+# IPython
+profile_default/
+ipython_config.py
+
+# pyenv
+#   For a library or package, you might want to ignore these files since the code is
+#   intended to run in multiple environments; otherwise, check them in:
+# .python-version
+
+# pipenv
+#   According to pypa/pipenv#598, it is recommended to include Pipfile.lock in version control.
+#   However, in case of collaboration, if having platform-specific dependencies or dependencies
+#   having no cross-platform support, pipenv may install dependencies that don't work, or not
+#   install all needed dependencies.
+#Pipfile.lock
+
+# PEP 582; used by e.g. github.com/David-OConnor/pyflow
+__pypackages__/
+
+# Celery stuff
+celerybeat-schedule
+celerybeat.pid
+
+# SageMath parsed files
+*.sage.py
+
+# Environments
+.env
+.venv
+env/
+venv/
+ENV/
+env.bak/
+venv.bak/
+
+# Spyder project settings
+.spyderproject
+.spyproject
+
+# Rope project settings
+.ropeproject
+
+# mkdocs documentation
+/site
+
+# mypy
+.mypy_cache/
+.dmypy.json
+dmypy.json
+
+# Pyre type checker
+.pyre/
+
+# pytype static type analyzer
+.pytype/
+
+# Cython debug symbols
+cython_debug/
+
+# cluster machine files
+mach*
+
+# Text-based output files
+*.csv
+
+# Exodus mesh files
+*.e
+
+# Griffin input files
+*.i
